+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.sungbin.kakaoemoticonparser">
+    package="me.sungbin.kakaoemoticonparser">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission
