--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/emoticon/EmoticonModule.kt
@@ -1,32 +1,0 @@
-package com.sungbin.kakaoemoticonparser.emoticon
-
-import dagger.Module
-import dagger.Provides
-import java.util.concurrent.TimeUnit
-import javax.inject.Singleton
-import okhttp3.OkHttpClient
-import retrofit2.Retrofit
-import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory
-import retrofit2.converter.gson.GsonConverterFactory
-
-@Module
-class EmoticonModule {
-
-    @Module
-    companion object {
-        private val defaultOkHttpClient = OkHttpClient.Builder()
-            .connectTimeout(30, TimeUnit.MINUTES)
-            .readTimeout(30, TimeUnit.SECONDS)
-            .writeTimeout(30, TimeUnit.SECONDS)
-
-        @JvmStatic
-        @Singleton
-        @Provides
-        fun provideSearchInstance() = Retrofit.Builder()
-            .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
-            .addConverterFactory(GsonConverterFactory.create())
-            .baseUrl("https://e.kakao.com/")
-            .client(defaultOkHttpClient.build())
-            .build()
-    }
-}
