--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/emoticon/model/detail/DetailBandBanner.kt
@@ -1,18 +1,0 @@
-package com.sungbin.kakaoemoticonparser.emoticon.model.detail
-
-import com.google.gson.annotations.SerializedName
-
-data class DetailBandBanner(
-
-    @field:SerializedName("targetMobileLinkUrl")
-    val targetMobileLinkUrl: String,
-
-    @field:SerializedName("targetPcLinkUrl")
-    val targetPcLinkUrl: String,
-
-    @field:SerializedName("imageUrl")
-    val imageUrl: String,
-
-    @field:SerializedName("bgColorCode")
-    val bgColorCode: String
-)
