--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/emoticon/model/detail/Response.kt
@@ -1,12 +1,0 @@
-package com.sungbin.kakaoemoticonparser.emoticon.model.detail
-
-import com.google.gson.annotations.SerializedName
-
-data class Response(
-
-    @field:SerializedName("result")
-    val result: Result,
-
-    @field:SerializedName("status")
-    val status: Int
-)
