--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/emoticon/model/detail/UsableTicket.kt
@@ -1,12 +1,0 @@
-package com.sungbin.kakaoemoticonparser.emoticon.model.detail
-
-import com.google.gson.annotations.SerializedName
-
-data class UsableTicket(
-
-    @field:SerializedName("ticketName")
-    val ticketName: String,
-
-    @field:SerializedName("saleFactor")
-    val saleFactor: Double
-)
