--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/theme/AppThemeState.kt
@@ -1,6 +1,0 @@
-package com.sungbin.kakaoemoticonparser.theme
-
-data class AppThemeState(
-    var darkTheme: Boolean = false,
-    var pallet: ColorPallet = ColorPallet.BLUE
-)
