--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/theme/Color.kt
@@ -1,21 +1,0 @@
-package com.sungbin.kakaoemoticonparser.theme
-
-import androidx.compose.ui.graphics.Color
-
-val green200 = Color(0xffa5d6a7)
-val green500 = Color(0xff4caf50)
-val green700 = Color(0xff388e3c)
-
-val blue500 = Color(0xFF3F51B5)
-val blue200 = Color(0xFF9FA8DA)
-val blue700 = Color(0xFF303F9F)
-
-val purple200 = Color(0xFFB39DDB)
-val purple = Color(0xFF833AB4)
-val purple700 = Color(0xFF512DA8)
-
-val orange200 = Color(0xFFff7961)
-val orange500 = Color(0xFFf44336)
-val orange700 = Color(0xFFba000d)
-
-val teal200 = Color(0xff80deea)
