--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/theme/Shape.kt
@@ -1,11 +1,0 @@
-package com.sungbin.kakaoemoticonparser.theme
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Shapes
-import androidx.compose.ui.unit.dp
-
-val shapes = Shapes(
-    small = RoundedCornerShape(4.dp),
-    medium = RoundedCornerShape(8.dp),
-    large = RoundedCornerShape(16.dp)
-)
