--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/theme/Theme.kt
@@ -1,134 +1,0 @@
-package com.sungbin.kakaoemoticonparser.theme
-
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.MaterialTheme.shapes
-import androidx.compose.material.darkColors
-import androidx.compose.material.lightColors
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
-
-private val DarkGreenColorPalette = darkColors(
-    primary = green200,
-    primaryVariant = green700,
-    secondary = teal200,
-    background = Color.Black,
-    surface = Color.Black,
-    onPrimary = Color.Black,
-    onSecondary = Color.White,
-    onBackground = Color.White,
-    onSurface = Color.White,
-    error = Color.Red,
-)
-
-private val DarkPurpleColorPalette = darkColors(
-    primary = purple200,
-    primaryVariant = purple700,
-    secondary = teal200,
-    background = Color.Black,
-    surface = Color.Black,
-    onPrimary = Color.Black,
-    onSecondary = Color.White,
-    onBackground = Color.White,
-    onSurface = Color.White,
-    error = Color.Red,
-)
-
-private val DarkBlueColorPalette = darkColors(
-    primary = blue200,
-    primaryVariant = blue700,
-    secondary = teal200,
-    background = Color.Black,
-    surface = Color.Black,
-    onPrimary = Color.Black,
-    onSecondary = Color.White,
-    onBackground = Color.White,
-    onSurface = Color.White,
-    error = Color.Red,
-)
-
-private val DarkOrangeColorPalette = darkColors(
-    primary = orange200,
-    primaryVariant = orange700,
-    secondary = teal200,
-    background = Color.Black,
-    surface = Color.Black,
-    onPrimary = Color.Black,
-    onSecondary = Color.White,
-    onBackground = Color.White,
-    onSurface = Color.White,
-    error = Color.Red,
-)
-
-private val LightGreenColorPalette = lightColors(
-    primary = green500,
-    primaryVariant = green700,
-    secondary = teal200,
-    background = Color.White,
-    surface = Color.White,
-    onPrimary = Color.White,
-    onSecondary = Color.Black,
-    onBackground = Color.Black,
-    onSurface = Color.Black
-)
-
-private val LightPurpleColorPalette = lightColors(
-    primary = purple,
-    primaryVariant = purple700,
-    secondary = teal200,
-    background = Color.White,
-    surface = Color.White,
-    onPrimary = Color.White,
-    onSecondary = Color.Black,
-    onBackground = Color.Black,
-    onSurface = Color.Black
-)
-
-private val LightBlueColorPalette = lightColors(
-    primary = blue500,
-    primaryVariant = blue700,
-    secondary = teal200,
-    background = Color.White,
-    surface = Color.White,
-    onPrimary = Color.White,
-    onSecondary = Color.Black,
-    onBackground = Color.Black,
-    onSurface = Color.Black
-)
-
-private val LightOrangeColorPalette = lightColors(
-    primary = orange500,
-    primaryVariant = orange700,
-    secondary = teal200,
-    background = Color.White,
-    surface = Color.White,
-    onPrimary = Color.White,
-    onSecondary = Color.Black,
-    onBackground = Color.Black,
-    onSurface = Color.Black
-)
-
-enum class ColorPallet {
-    PURPLE, GREEN, ORANGE, BLUE
-}
-
-@Composable
-fun AppTheme(
-    darkTheme: Boolean = isSystemInDarkTheme(),
-    colorPallet: ColorPallet = ColorPallet.BLUE,
-    content: @Composable () -> Unit,
-) {
-    val colors = when (colorPallet) {
-        ColorPallet.GREEN -> if (darkTheme) DarkGreenColorPalette else LightGreenColorPalette
-        ColorPallet.PURPLE -> if (darkTheme) DarkPurpleColorPalette else LightPurpleColorPalette
-        ColorPallet.ORANGE -> if (darkTheme) DarkOrangeColorPalette else LightOrangeColorPalette
-        ColorPallet.BLUE -> if (darkTheme) DarkBlueColorPalette else LightBlueColorPalette
-    }
-
-    MaterialTheme(
-        colors = colors,
-        typography = typography,
-        shapes = shapes,
-        content = content
-    )
-}
