--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/ui/NavigationType.kt
@@ -1,7 +1,0 @@
-package com.sungbin.kakaoemoticonparser.ui
-
-enum class NavigationType {
-    SEARCH,
-    FAVORITE,
-    SETTING
-}
