--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/ui/search/SearchContentState.kt
@@ -1,8 +1,0 @@
-package com.sungbin.kakaoemoticonparser.ui.search
-
-enum class SearchContentState {
-    HOME,
-    RESULT,
-    NULL, // 검색결과 없음
-    ERROR
-}
