--- app/src/main/kotlin/com/sungbin/kakaoemoticonparser/ui/test/TestContent.kt
@@ -1,11 +1,0 @@
-package com.sungbin.kakaoemoticonparser.ui.test
-
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.res.stringResource
-import com.sungbin.kakaoemoticonparser.R
-
-@Composable
-fun TestContent() {
-    Text(text = stringResource(id = R.string.app_name))
-}
