+++ app/src/main/kotlin/me/sungbin/kakaoemoticonparser/emoticon/model/Result.kt
@@ -1,0 +1,15 @@
+package me.sungbin.kakaoemoticonparser.emoticon.model
+
+import com.google.gson.annotations.SerializedName
+
+data class Result(
+
+    @field:SerializedName("last")
+    val last: Boolean,
+
+    @field:SerializedName("totalCount")
+    val totalCount: Int,
+
+    @field:SerializedName("content")
+    val content: List<ContentItem>
+)
