+++ app/src/main/kotlin/me/sungbin/kakaoemoticonparser/theme/AppThemeState.kt
@@ -1,0 +1,6 @@
+package me.sungbin.kakaoemoticonparser.theme
+
+data class AppThemeState(
+    var darkTheme: Boolean = false,
+    var pallet: ColorPallet = ColorPallet.BLUE
+)
