+++ app/src/main/kotlin/me/sungbin/kakaoemoticonparser/theme/SystemUiController.kt
@@ -1,0 +1,63 @@
+package me.sungbin.kakaoemoticonparser.theme
+
+import android.os.Build
+import android.view.View
+import android.view.Window
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.compositeOver
+import androidx.compose.ui.graphics.luminance
+import androidx.compose.ui.graphics.toArgb
+
+@Suppress("DEPRECATION")
+class SystemUiController(private val window: Window) {
+    private val blackScrim = Color(0f, 0f, 0f, 0.2f)
+    private val blackScrimmed: (Color) -> Color = { original ->
+        blackScrim.compositeOver(original)
+    }
+
+    fun setStatusBarColor(
+        color: Color,
+        darkIcons: Boolean = color.luminance() > 0.5f
+    ) {
+        window.statusBarColor = color.toArgb()
+        if (darkIcons) {
+            window.decorView.systemUiVisibility = window.decorView.systemUiVisibility or
+                View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR
+        } else {
+            window.decorView.systemUiVisibility = window.decorView.systemUiVisibility and
+                View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR.inv()
+        }
+    }
+
+    fun setNavigationBarColor(
+        color: Color,
+        darkIcons: Boolean = color.luminance() > 0.5f,
+        transformColorForLightContent: (Color) -> Color = blackScrimmed
+    ) {
+        val navBarColor = when {
+            Build.VERSION.SDK_INT >= 29 -> Color.Transparent
+            darkIcons && Build.VERSION.SDK_INT < 26 -> transformColorForLightContent(color)
+            else -> color
+        }
+        window.navigationBarColor = navBarColor.toArgb()
+        if (Build.VERSION.SDK_INT >= 26) {
+            @Suppress("DEPRECATION")
+            if (darkIcons) {
+                window.decorView.systemUiVisibility = window.decorView.systemUiVisibility or
+                    View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR
+            } else {
+                window.decorView.systemUiVisibility = window.decorView.systemUiVisibility and
+                    View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR.inv()
+            }
+        }
+    }
+
+    fun setSystemBarsColor(
+        color: Color,
+        darkIcons: Boolean = color.luminance() > 0.5f,
+        transformColorForLightContent: (Color) -> Color = blackScrimmed
+    ) {
+        setStatusBarColor(color, darkIcons)
+        setNavigationBarColor(color, darkIcons, transformColorForLightContent)
+    }
+}
