+++ app/src/main/kotlin/me/sungbin/kakaoemoticonparser/ui/NavigationType.kt
@@ -1,0 +1,7 @@
+package me.sungbin.kakaoemoticonparser.ui
+
+enum class NavigationType {
+    SEARCH,
+    FAVORITE,
+    SETTING
+}
