+++ app/src/main/kotlin/me/sungbin/kakaoemoticonparser/ui/search/SearchContentState.kt
@@ -1,0 +1,8 @@
+package me.sungbin.kakaoemoticonparser.ui.search
+
+enum class SearchContentState {
+    HOME,
+    RESULT,
+    NULL, // 검색결과 없음
+    ERROR
+}
