+++ app/src/main/kotlin/me/sungbin/kakaoemoticonparser/ui/test/TestContent.kt
@@ -1,0 +1,11 @@
+package me.sungbin.kakaoemoticonparser.ui.test
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import me.sungbin.kakaoemoticonparser.R
+
+@Composable
+fun TestContent() {
+    Text(text = stringResource(id = R.string.app_name))
+}
